Add tempo snackbar host and review snackbar style
