+++ app/src/main/java/com/enricog/tempo/MainActivity.kt
@@ -4,7 +4,6 @@ import android.os.Bundle
 import android.view.WindowManager
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.material.Scaffold
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.lifecycle.lifecycleScope
@@ -68,8 +67,6 @@ internal class MainActivity : ComponentActivity() {
                     }
                 }
 
-                Scaffold() {
-                }
                 ModalBottomSheetLayout(
                     bottomSheetNavigator = bottomSheetNavigator,
                     sheetShape = TempoTheme.shapes.bottomSheet,
