+++ features/routines/src/main/java/com/enricog/features/routines/list/ui_components/RoutinesScene.kt
@@ -1,9 +1,15 @@
 package com.enricog.features.routines.list.ui_components
 
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
@@ -13,7 +19,10 @@ import com.enricog.data.routines.api.entities.Routine
 import com.enricog.features.routines.R
 import com.enricog.ui.components.button.TempoButtonColor
 import com.enricog.ui.components.button.icon.TempoIconButton
+import com.enricog.ui.components.snackbar.TempoSnackbarHost
+import com.enricog.ui.components.snackbar.rememberSnackbarHostState
 import com.enricog.ui.theme.TempoTheme
+import kotlinx.coroutines.launch
 
 internal const val RoutinesSceneTestTag = "RoutinesSceneTestTag"
 
@@ -24,6 +33,10 @@ internal fun RoutinesScene(
     onRoutineDelete: (Routine) -> Unit,
     onCreateRoutineClick: () -> Unit,
 ) {
+
+    val snackbarHostState = rememberSnackbarHostState()
+    val scope = rememberCoroutineScope()
+
     Box(
         modifier = Modifier
             .testTag(RoutinesSceneTestTag)
@@ -48,11 +61,17 @@ internal fun RoutinesScene(
             }
         }
 
+        TempoSnackbarHost(state = snackbarHostState)
+
         TempoIconButton(
             modifier = Modifier
                 .align(Alignment.BottomEnd)
                 .padding(TempoTheme.dimensions.spaceL),
-            onClick = onCreateRoutineClick,
+            onClick = {
+                scope.launch {
+                    snackbarHostState.show("test")
+                }
+            },
             icon = painterResource(R.drawable.ic_add),
             color = TempoButtonColor.Accent,
             contentDescription = stringResource(R.string.content_description_button_create_routine)
